+++ pkg/grpc/email_client_emulator/send_email.go
@@ -3,7 +3,9 @@ package email_client_emulator
 import (
 	"bufio"
 	"context"
+	"errors"
 	"os"
+	"strconv"
 	"time"
 
 	"github.com/coneno/logger"
@@ -27,14 +29,21 @@ func (s *emailClientServer) SendEmail(ctx context.Context, req *api.SendEmailReq
 	}
 
 	var err error
+	var fileCounter = 1
 	for _, to := range req.To {
 		filepath := s.EmailClientEmulatorPath + "/" + to
 		err = os.MkdirAll(filepath, os.ModePerm)
 		if err != nil {
 			logger.Error.Printf("error sending mail: err at target path mkdir %v", err.Error())
 		}
-		filename := time.Now().Format("2006-01-02 15:04:05") + " " + req.Subject + ".html"
-		f, err := os.Create(filepath + "/" + filename)
+		filename := time.Now().Format("2006-01-01 15:04:05") + " " + req.Subject
+		originFilename := filename
+		fileType := ".html"
+		for CheckIfFileExits(filepath + "/" + filename + fileType) {
+			fileCounter++
+			filename = originFilename + "(" + strconv.Itoa(fileCounter) + ")"
+		}
+		f, err := os.Create(filepath + "/" + filename + fileType)
 		if err != nil {
 			logger.Error.Printf("error while creating file %v", filename)
 		}
@@ -54,4 +63,9 @@ func (s *emailClientServer) SendEmail(ctx context.Context, req *api.SendEmailReq
 		Status:  api.ServiceStatus_NORMAL,
 		Msg:     "email sent",
 	}, nil
+}
+
+func CheckIfFileExits(filepath string) bool {
+	_, error := os.Stat(filepath)
+	return !errors.Is(error, os.ErrNotExist)
 }