+++ pkg/templates/templates.go
@@ -5,6 +5,7 @@ import (
 	"encoding/base64"
 	"errors"
 	"html/template"
+	"strings"
 
 	"github.com/coneno/logger"
 	"github.com/influenzanet/messaging-service/pkg/types"
@@ -23,6 +24,10 @@ func GetTemplateTranslation(tDef types.EmailTemplate, lang string) types.Localiz
 }
 
 func ResolveTemplate(tempName string, templateDef string, contentInfos map[string]string) (content string, err error) {
+	if strings.TrimSpace(templateDef) == "" {
+		logger.Error.Printf("error: empty template %s", tempName)
+		return "", errors.New("empty template `" + tempName)
+	}
 	tmpl, err := template.New(tempName).Parse(templateDef)
 	if err != nil {
 		logger.Error.Printf("error when parsing template %s: %v", tempName, err)
@@ -39,7 +44,10 @@ func ResolveTemplate(tempName string, templateDef string, contentInfos map[strin
 }
 
 func CheckAllTranslationsParsable(tempTranslations types.EmailTemplate) (err error) {
-
+	if len(tempTranslations.Translations) == 0 {
+		logger.Error.Printf("error when decoding template %s: translation list is empty", tempTranslations.MessageType)
+		return errors.New("error when decoding template `" + tempTranslations.MessageType + "`: translation list is empty")
+	}
 	for _, templ := range tempTranslations.Translations {
 		templateName := tempTranslations.MessageType + templ.Lang
 		decodedTemplate, err := base64.StdEncoding.DecodeString(templ.TemplateDef)
@@ -56,6 +64,5 @@ func CheckAllTranslationsParsable(tempTranslations types.EmailTemplate) (err err
 			return errors.New("could not parse template for `" + templ.Lang + "` - error: " + err.Error())
 		}
 	}
-
 	return nil
 }