+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 
+- [PR #12](https://github.com/influenzanet/messaging-service/pull/12) Scheduled messages for participants are now fetched and sent  regularly with specified frequency for all studies. Researcher notifications are also fetched and sent with specified frequency. Uploading auto message schedules for these message types is omitted.
 - [PR #10](https://github.com/influenzanet/messaging-service/pull/10) Checks if go template engine is able to parse and execute new email templates. This prevents uploading incorrect email templates.
 - [PR #11](https://github.com/influenzanet/messaging-service/pull/11) Improvement on log messages and raise error when no instances are found to help discover issues during initial setup.